+++ fif/main.py
@@ -2,37 +2,47 @@ import sys
 import argparse
 import os
 import time
-import filetype 
+import filetype
 from fif import gif_codec
 
 sys.tracebacklimit = 0
 
-# SUPPORTED_EXT = [png]
 
 def getargs():
     parser = argparse.ArgumentParser()
-    parser.add_argument("-e", "--encode", help="encode file", type=str, choices=("png", "wav","gif"))
-    parser.add_argument("-v", "--verbose", action="store_true", help="Increase output verbosity")
+    parser.add_argument(
+        "-e",
+        "--encode",
+        help="encode file",
+        type=str,
+        choices=(
+            "png",
+            "wav",
+            "gif"))
+    parser.add_argument(
+        "-v",
+        "--verbose",
+        action="store_true",
+        help="Increase output verbosity")
     parser.add_argument("input", type=str, help="Input file path")
     return parser.parse_args()
 
+
 def run_fif():
     args = getargs()
-    # print(args)
-    # print(filetype.TYPES)
-    if args.encode == None:
-        # decoding
-        kind = filetype.guess(args.input) # https://github.com/h2non/filetype.py#supported-types
+    if args.encode is None:
+        kind = filetype.guess(args.input)
         if kind.mime == "image/gif":
             if args.verbose:
                 print('Container file extension: %s' % kind.extension)
                 print('Container MIME type: %s' % kind.mime)
-            gif_codec.decode(args.input, verbose=args.verbose)     
+            gif_codec.decode(args.input, verbose=args.verbose)
         elif kind.mime == "image/png":
             print("png decoding coming soon!")
         elif kind.mime == "audio/x-wav":
             print("wav decoding coming soon!")
-            
+        else:
+            print("Unsupported MIME")
     if args.encode == "gif":
         gif_codec.encode(args.input, verbose=args.verbose)
     elif args.encode == "png":