+++ CHANGELOG.md
@@ -3,6 +3,12 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [1.0.0.dev1] - 2020-11-25
+### Changed
+- The output file is saved in the same folder as the input file.
+- Now, there is no byte difference between the embedded file and the extracted file.
+- Code cleaning.
+
 ## [0.1.0.dev1] - 2020-11-24
 ### Added
 - Encoding and decoding grayscale GIF (L mode)